fix(qty): kuantitas desimal utuh di respons + pratinjau + pembulatan bill HALF_UP
Unit KUANTITAS DESIMAL (P0 + item4 + pembulatan), atas putusan MASTER.
Gerbang dua sisi 12/12 GREEN (schema + BillCalculator, OLD vs NEW in-container).

Cacat: kuantitas pecahan (mis. 0,51) terpotong jadi 0 di jalur BACA, walau DB
menyimpan numeric(15,4) dengan benar (terukur: bill_items.quantity=0.5100).

P0 — hapus pemotong int() kuantitas:
- schemas/bills.py: BillItemResponseV2.qty int -> float (Law 25: respons=float).
- services/bills_service.py: 2x "qty": int(item["quantity"]) -> float (pembangun
  respons V1/V2). WAJIB seiring schema: Pydantic v2 qty:int MENOLAK 0.51
  (int_from_float), jadi fix schema tanpa fix service -> 500.
- routers/stock_adjustments.py: 5 int() kuantitas (quantity_before/adjustment/
  after + system/physical_quantity) -> float. Komentar lama "Law 25: read path"
  keliru pakai int; Law 25 = respons int->FLOAT, bukan int.

Item4 — pratinjau jurnal = posting:
- routers/bills.py preview_journal: int(grand_total)/int(tax) memangkas sen ->
  round(...,2). Pratinjau yang beda dari posting berbohong.

Pembulatan bill -> HALF_UP per-baris (satu aturan dgn faktur penjualan):
- BillCalculator.calculate: dulu jumlah baris tak dibulatkan lalu total quantize
  HALF_EVEN; kini per-baris quantize(0.01, ROUND_HALF_UP), sama seperti
  sales_invoices._r2. Semua quantize(TWO) -> ROUND_HALF_UP eksplisit.
- Dampak terukur ke bill lama: 0 dari 45 bill berubah (semua qty lama integer).
  Forward-only; nol jurnal POSTED disentuh (Law 19).

Ukur pendukung: nol dokumen riil punya kuantitas pecahan selain 2 baris
warehouse_stock (MATERIAL_ISSUE WO-2026-000034, produksi nyata, POSTED).

Backlog terpisah (perintah MASTER, TAK di commit ini): int() pada AMOUNT jalur
tulis purchase_orders.py:1270/1332 & recurring_invoices.py:247 (PO->bill,
template berulang) -- ukur dulu. Lalu P1: float->Decimal request quantity
(PO/quote/SO/CN/vendor_credit/stock_adj) + action_chat qty.

+++ backend/api_gateway/app/routers/stock_adjustments.py
@@ -417,23 +417,17 @@ async def get_stock_adjustment(request: Request, adjustment_id: UUID):
                             "product_id": str(item["product_id"]),
                             "product_code": item["product_code"],
                             "product_name": item["product_name"],
-                            "quantity_before": int(
-                                item["quantity_before"]
-                            ),  # Law 25: read path
-                            "quantity_adjustment": int(
-                                item["quantity_adjustment"]
-                            ),  # Law 25: read path
-                            "quantity_after": int(
-                                item["quantity_after"]
-                            ),  # Law 25: read path
+                            "quantity_before": float(item["quantity_before"]),  # Law 25: read path
+                            "quantity_adjustment": float(item["quantity_adjustment"]),  # Law 25: read path
+                            "quantity_after": float(item["quantity_after"]),  # Law 25: read path
                             "unit": item["unit"],
                             "unit_cost": item["unit_cost"],
                             "total_value": item["total_value"],
                             "reason_detail": item["reason_detail"],
-                            "system_quantity": int(item["system_quantity"])
+                            "system_quantity": float(item["system_quantity"])
                             if item["system_quantity"]
                             else None,  # Law 25: read path
-                            "physical_quantity": int(item["physical_quantity"])
+                            "physical_quantity": float(item["physical_quantity"])
                             if item["physical_quantity"]
                             else None,  # Law 25: read path
                             "line_number": item["line_number"],
